+++ src/platform/src/Contract/Normalizer/Message/Content/DocumentUrlNormalizer.php
@@ -9,38 +9,38 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\AI\Platform\Contract\Normalizer\Message;
+namespace Symfony\AI\Platform\Contract\Normalizer\Message\Content;
 
-use Symfony\AI\Platform\Message\SystemMessage;
+use Symfony\AI\Platform\Message\Content\DocumentUrl;
 use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
 
 /**
- * @author Christopher Hertel <mail@christopher-hertel.de>
+ * @author Guillaume Loulier <personal@guillaumeloulier.fr>
  */
-final class SystemMessageNormalizer implements NormalizerInterface
+final class DocumentUrlNormalizer implements NormalizerInterface
 {
     public function supportsNormalization(mixed $data, ?string $format = null, array $context = []): bool
     {
-        return $data instanceof SystemMessage;
+        return $data instanceof DocumentUrl;
     }
 
     public function getSupportedTypes(?string $format): array
     {
         return [
-            SystemMessage::class => true,
+            DocumentUrl::class => true,
         ];
     }
 
     /**
-     * @param SystemMessage $data
+     * @param DocumentUrl $data
      *
-     * @return array{role: 'system', content: string}
+     * @return array{type: 'document_url', document_url: array{url: string}}
      */
     public function normalize(mixed $data, ?string $format = null, array $context = []): array
     {
         return [
-            'role' => $data->getRole()->value,
-            'content' => (string) $data->getContent(),
+            'type' => 'document_url',
+            'document_url' => ['url' => $data->getUrl()],
         ];
     }
 }